+++ src/server/devices/groups/index.ts
@@ -1,4 +1,3 @@
-export * from './acGroup';
 export * from './base-group';
 export * from './Fenster';
 export * from './fensterGroup';